convrted colin-api structure to follow our other services
