+++ colin-api/requirements.txt
@@ -13,9 +13,8 @@ cx-Oracle==7.3.0
 ecdsa==0.14.1
 flask-jwt-oidc==0.1.5
 flask-restplus==0.13.0
-future==0.18.2
 gunicorn==20.0.4
-importlib-metadata==1.2.0
+importlib-metadata==1.3.0
 itsdangerous==1.1.0
 jsonschema==3.2.0
 more-itertools==8.0.2
@@ -24,7 +23,7 @@ pyasn1==0.4.8
 pycountry==19.8.18
 pyrsistent==0.15.6
 python-dotenv==0.10.3
-python-jose==3.0.1
+python-jose==3.1.0
 pytz==2019.3
 rsa==4.0
 sentry-sdk==0.13.5
