+++ colin-api/setup.cfg
@@ -46,7 +46,7 @@ max_line_length = 120
 ignore = E501
 docstring-min-length=10
 notes=FIXME,XXX # TODO is ignored
-match_dir = colin_api
+match_dir = src/colin_api
 ignored-modules=flask_sqlalchemy
     sqlalchemy
 per-file-ignores =
@@ -95,10 +95,10 @@ markers =
 [coverage:run]
 branch = True
 source =
-    colin_api
+    src/colin_api
 omit =
-    colin_api/wsgi.py
-    colin_api/gunicorn_config.py
+    src/colin_api/wsgi.py
+    src/colin_api/gunicorn_config.py
 
 [report:run]
 exclude_lines =
