+++ colin-api/src/colin_api/__init__.py
@@ -23,9 +23,9 @@ from flask import Flask
 from flask_jwt_oidc import JwtManager
 
 import config
-from colin_api.resources import API_BLUEPRINT, OPS_BLUEPRINT
-from colin_api.utils.logging import setup_logging
-from colin_api.utils.run_version import get_run_version
+from src.colin_api.resources import API_BLUEPRINT, OPS_BLUEPRINT
+from src.colin_api.utils.logging import setup_logging
+from src.colin_api.utils.run_version import get_run_version
 # noqa: I003; the sentry import creates a bad line count in isort
 
 setup_logging(os.path.join(os.path.abspath(os.path.dirname(__file__)), 'logging.conf'))  # important to do this first
@@ -46,7 +46,7 @@ def create_app(run_mode=os.getenv('FLASK_ENV', 'production')):
         )
     app.register_blueprint(API_BLUEPRINT)
     app.register_blueprint(OPS_BLUEPRINT)
-    setup_jwt_manager(app, jwt)
+    # setup_jwt_manager(app, jwt)
 
     @app.after_request
     def add_version(response):  # pylint: disable=unused-variable
