+++ colin-api/src/colin_api/exceptions/__init__.py
