+++ colin-api/src/colin_api/models/business.py
@@ -19,9 +19,9 @@ from datetime import datetime
 
 from flask import current_app
 
-from colin_api.exceptions import BusinessNotFoundException
-from colin_api.resources.db import DB
-from colin_api.utils import convert_to_json_date, convert_to_json_datetime
+from src.colin_api.exceptions import BusinessNotFoundException
+from src.colin_api.resources.db import DB
+from src.colin_api.utils import convert_to_json_date, convert_to_json_datetime
 
 
 class Business:
