+++ colin-api/src/colin_api/resources/meta.py
@@ -11,19 +11,25 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Supply version and commit hash info."""
-import os
+"""Meta information about the service.
 
-from colin_api.version import __version__
+Currently this only provides API versioning information
+"""
+from flask import jsonify
+from flask_restplus import Namespace, Resource
+
+from src.colin_api.utils.run_version import get_run_version
 
 
-def _get_build_openshift_commit_hash():
-    return os.getenv('OPENSHIFT_BUILD_COMMIT', None)
+API = Namespace('Meta', description='Metadata')
 
 
-def get_run_version():
-    """Return a formatted version string for this service."""
-    commit_hash = _get_build_openshift_commit_hash()
-    if commit_hash:
-        return f'{__version__}-{commit_hash}'
-    return __version__
+@API.route('/info')
+class Info(Resource):
+    """Meta information about the overall service."""
+
+    @staticmethod
+    def get():
+        """Return a JSON object with meta information about the Service."""
+        version = get_run_version()
+        return jsonify(API=f'colin_api/{version}')
