+++ colin-api/src/colin_api/resources/office.py
@@ -19,10 +19,10 @@ Currently this only provides API versioning information
 from flask import current_app, jsonify
 from flask_restplus import Resource, cors
 
-from colin_api.exceptions import GenericException
-from colin_api.models import Office
-from colin_api.resources.business import API
-from colin_api.utils.util import cors_preflight
+from src.colin_api.exceptions import GenericException
+from src.colin_api.models import Office
+from src.colin_api.resources.business import API
+from src.colin_api.utils.util import cors_preflight
 
 
 @cors_preflight('GET')
