+++ colin-api/src/colin_api/resources/ops.py
@@ -16,7 +16,7 @@ import cx_Oracle
 from flask import current_app
 from flask_restplus import Namespace, Resource
 
-from colin_api.resources.db import DB
+from src.colin_api.resources.db import DB
 
 
 API = Namespace('OPS', description='Service - OPS checks')
