+++ colin-api/tests/unit/api/test_meta.py
@@ -22,7 +22,7 @@ from tests import skip_in_pod
 @skip_in_pod
 def test_meta_no_commit_hash(client):
     """Assert that the endpoint returns just the services __version__."""
-    from colin_api.version import __version__
+    from src.colin_api.version import __version__
 
     rv = client.get('/api/v1/meta/info')
 
@@ -32,7 +32,7 @@ def test_meta_no_commit_hash(client):
 
 def test_meta_with_commit_hash(monkeypatch, client):
     """Assert that the endpoint return __version__ and the last git hash used to build the services image."""
-    from colin_api.version import __version__
+    from src.colin_api.version import __version__
 
     commit_hash = 'deadbeef_ha'
     monkeypatch.setenv('OPENSHIFT_BUILD_COMMIT', commit_hash)
