+++ colin-api/tests/unit/conf/test_configuration.py
@@ -21,7 +21,6 @@ import pytest
 
 import config
 
-
 # testdata pattern is ({str: environment}, {expected return value})
 TEST_ENVIRONMENT_DATA = [
     ('valid', 'development', config.DevConfig),
