+++ colin-api/tests/unit/conf/test_version.py
@@ -16,8 +16,8 @@
 
 Test-Suite to ensure that the version utilities are working as expected.
 """
-from colin_api import utils
-from colin_api.version import __version__
+from src.colin_api import utils
+from src.colin_api.version import __version__
 from tests import skip_in_pod
